Ten commits after linux mint 19.3 upgrade

+++ src/main/java/org/springframework/samples/escalade/repository/LengthRepository.java
@@ -5,7 +5,7 @@ import org.springframework.samples.escalade.model.Length;
 
 public interface LengthRepository {
 
-	Length findById(Long id);
+	Length findById(Integer id);
 	void saveLength(Length length)throws DataAccessException;
 
 }
