+++ src/main/java/org/springframework/samples/escalade/repository/WayRepository.java
@@ -5,8 +5,8 @@ import org.springframework.samples.escalade.model.Way;
 
 public interface WayRepository {
 
-	Way findById(Long id) throws DataAccessException;
+	Way findById(Integer id) throws DataAccessException;
 
 	void saveWay(Way way) throws DataAccessException;
-
+	
 }
